temporary fix #175
Fixes the boat sinking issue.
When the server is restarted or the world is unloaded, the boat will start to sink again. As a workaround, when you try to drive the boat that started to sink, the boat will fall into position as an item, and you can drive without any problems when you put it back on. This is only triggered when the entity is regenerating.